+++ src/main/java/roomescape/controller/UserPageController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.user;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
