--- src/main/java/roomescape/controller/user/ThemeController.java
@@ -1,12 +1,0 @@
-package roomescape.controller.user;
-
-//@RestController
-//@RequestMapping("/themes")
-public class ThemeController {
-
-//    @GetMapping
-//    public ResponseEntity<List<ThemeResponseDto>> findAll() {
-//        return ResponseEntity.ok(themeService.findAll());
-//    }
-
-}
